fix double precision issues

+++ cheetah/accelerator/transverse_deflecting_cavity.py
@@ -15,7 +15,8 @@ from .element import Element
 generate_unique_name = UniqueNameGenerator(prefix="unnamed_element")
 
 electron_mass_eV = torch.tensor(
-    physical_constants["electron mass energy equivalent in MeV"][0] * 1e6
+    physical_constants["electron mass energy equivalent in MeV"][0] * 1e6,
+    dtype=torch.float64,
 )
